🧹 [code health improvement] Refactor WebcamAvatar into modular sub-components
🎯 **What:** Decoupled the 600+ line monolith `WebcamAvatar.tsx` into modular components: `WebcamOverlay`, `MannequinPreview`, and extracted Three.js initialization into `initThreeRig.ts`. Extracted base types into `types.ts`.
💡 **Why:** Significantly improves maintainability and readability by separating UI concerns from Three.js scene setup.
✅ **Verification:** Ran TypeScript compilation (tsc), formatting (prettier), and Vite build without errors. Ensured that real-time biometric mapping remains imperative (via mutable refs) instead of React state, preserving the 22ms latency constraint.
✨ **Result:** A cleaner component tree and pure logic extraction that protects the strict real-time performance architectural rules.
